+++ src/renderer/features/multisig-operations/model/context.ts
@@ -1,5 +1,5 @@
 import { combine, createEvent, restore, sample } from 'effector';
-import { throttle } from 'patronum';
+import { interval, throttle } from 'patronum';
 
 import { TransactionType } from '@/shared/core';
 import { type AnyAccount, type MultisigOperation, accountService, multisigOperation } from '@/domains/network';
@@ -98,6 +98,17 @@ sample({
 sample({
   clock: throttle(multisigOperationsFeature.stopped, 500),
   target: multisigOperation.unsubscribeFromAccounts,
+});
+
+const { tick } = interval({
+  start: multisigOperationsFeature.running,
+  stop: multisigOperationsFeature.stopped,
+  timeout: 30000,
+});
+
+sample({
+  clock: tick,
+  target: multisigOperation.refetchOffchainOperations,
 });
 
 export const operationsContextModel = {
